update document replacement
